Moved NumberShower to its own Storybook, and slight cleanup to other code.
