+++ packages/components/src/stories/Introduction.stories.mdx
@@ -2,8 +2,5 @@ import { Meta } from "@storybook/addon-docs";
 
 <Meta title="Squiggle/Introduction" />
 
-This is the component library for Squiggle. All of these components are react
+This is the component library for Squiggle. These are React
 components, and can be used in any application that you see fit.
-
-Currently, the only component that is provided is the SquiggleChart component.
-This component allows you to render the result of a squiggle expression.
